backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pillow==10.4.0
python-multipart==0.0.9
